// src/types/IMailParams.ts
export interface IMailParams {
    transactionNumber: string;
    dateAndTIme: string;
    etherscanLink: string;
    firstName: string;
    lastName: string;
    email: string;
    productName: string;
    productPrice: string;
    quantity: number;
    amountReceived: number;
    coinSymbol: string;
}